+++ app/models/fae/static_page.rb
@@ -30,16 +30,27 @@ module Fae
 
       fae_fields.each do |name, value|
         type = value.is_a?(Hash) ? value[:type] : value
+        language = value.try(:[], :language)
 
-        define_association(name, type)
+        if language.present?
+          language.each do |lang|
+            define_association("#{name}_#{lang}", type) # Save with suffix for form fields
+          end
+          define_association(name, type, "#{name}_#{I18n.locale}") # Save with lookup to have default language return in front-end use
+        else
+          define_association(name, type) # Everything else
+        end
+
         define_validations(name, type, value[:validates]) if value.is_a?(Hash) && value[:validates].present?
       end
 
       @singleton_is_setup = true
     end
 
-    def self.define_association(name, type)
-      send :has_one, name.to_sym, -> { where(attached_as: name.to_s)}, as: poly_sym(type), class_name: type.to_s, dependent: :destroy
+    def self.define_association(name, type, locale_name = nil)
+      locale_name ||= name
+
+      send :has_one, name.to_sym, -> { where(attached_as: locale_name.to_s)}, as: poly_sym(type), class_name: type.to_s, dependent: :destroy
       send :accepts_nested_attributes_for, name, allow_destroy: true
       send :define_method, :"#{name}_content", -> { send(name.to_sym).try(:content) }
     end